// src/app/api-client/model/structureDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CandidateDto } from './candidateDto';


export interface StructureDto { 
    id?: number;
    structureName?: string;
    capacity?: number;
    adresse?: string;
    mobile?: string;
    statut?: string;
    log?: number;
    lat?: number;
    cityId?: number;
    createdAt?: string;
    avantages?: string;
    description?: string;
    images?: Array<string>;
    candidates?: Set<CandidateDto>;
}